requirements.txt
# OAuth for Dummies - Dependencies
# Install: pip install -r requirements.txt

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
